feat: 发布 v1.2.0 - 添加 32 位架构支持和 Windows 7 兼容性
🚀 主要新功能:
- 新增 32 位架构支持 (Windows 386, Linux 386, Linux ARM)
- Go 1.20 兼容性 (支持 Windows 7 系统)
- 编译路径隐私保护 (-trimpath 标志)

🏗️ 架构支持:
- Windows: amd64 (64位), 386 (32位), arm64
- Linux: amd64 (64位), 386 (32位), arm64, arm (32位)
- macOS: amd64 (Intel), arm64 (Apple Silicon)
- 总计 9 个不同平台和架构

🔧 技术改进:
- 降级 Go 版本从 1.24 到 1.20 以支持旧系统
- 降级依赖项版本确保兼容性
- 修复 32 位系统 syscall.Stat_t 类型转换问题
- 更新构建脚本支持所有新架构

📝 文档更新:
- 更新 README.md 和 BUILD_GUIDE.md
- 新增 build-32bit Makefile 目标
- 更新 GitHub Actions 支持 32 位构建

🛡️ 隐私保护:
- 所有构建使用 -trimpath 标志
- 移除编译后二进制文件中的本地路径信息

版本: v1.2.0

+++ internal/collectors/filesystem_linux.go
@@ -12,7 +12,7 @@ import (
 // getFileAccessTime 获取文件访问时间 (Linux)
 func getFileAccessTime(info os.FileInfo) time.Time {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return time.Unix(stat.Atim.Sec, stat.Atim.Nsec)
+		return time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec))
 	}
 	return info.ModTime()
 }
